feat: order repository

+++ Buy-NET.API/Domain/Models/OrderItem.cs
@@ -5,11 +5,11 @@ namespace Buy_NET.API.Domain.Models;
 public class OrderItem
 {
     [Key]
-    public int Id { get; set; }
+    public long Id { get; set; }
     [Required]
-    public int OrderId { get; set; }
+    public long OrderId { get; set; }
     [Required]
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
     [Required(ErrorMessage = "É necessário informar a quantidadade")]
     public int Quantity { get; set; }
     public Order Order { get; set; }
